NLP project I did today, first NLP project ever.
Took about 3 hours, tried to do a bit of customization but since it is my first time doing NLP I didn't really have much to add. Program is pretty decent though the text generation is not amazing lol Not much I could do about that, definitely fun to play with and has a lot of potential for future updates as well as being a good example of how quickly I am progressing on multiple fronts. Through if an H.P Lovecraft and Hunter S. Thompson book as text data to make it a bit more amusing. (On top of the given game of thrones text data)