New project finished
New project finished, a terminal game involving robots, exploring locations and collecting "titanium" as points and to buy items (project name was Duskers on hyperskill). 
The project is not fully finished yet (or at least what I plan to do with it, actual lesson/project itself is done), don't have the time to finish it as I have a lot to learn before I start my university classes in sept.